The new blameless postmortem...
After running a blameless postmortem last week, I wanted to take the opportunity to add this to the culture repo for easier access and update some of the language, as it felt a little harsh given what we are trying to achieve. 

I've also condensed some of the points of discussion as the format felt a little repetitive in places. Hopefully everyone feels we'll still get the same information out of the headers that I've changed to. 

Wanted to get everyone's opinion, rather than just putting it straight into the repo.