Create Kick Off Checklist
From our experience working on launching new products we've developed a series of sessions that need to happen before we begin work. I thought it was useful to share this with the wider web team as there may be others who benefit from this framework. Open to feedback & suggestions - very much still a work in progress (will be put to test again soon with A2B overseas transfers integration).